Raw Sql query
future work(or someone help me to do this):
The problem now is, all the current time is not set to a hard code
time(ex. 2001-01-01 01:00:01 UDT). Program fail to use Time.now(),Now().
etc. If we want to get dynamic time, we should fix this problem.
Otherwise we can ignore this problem